feat(dfu): refactor to modules
